Update pop_projections.R
Hi Matt! I noticed that your file doesn't check for packages, so it cannot be run off the bat for the average user so I added a check to the top. Also, I would recommend using the XLConnect package rather than the readxl package, it's considerably more functional and is platform-agnostic.